reorganization of accounts file and configuration modules
