+++ src/nodejs/get_worth.js
@@ -2,9 +2,9 @@
 
 const dbg     = require('./debug')
 const program = require('commander')
-var cfg       = require('./neoscan/config.js')
+var cfg       = require('./config.js')
 const _       = require('underscore')
-var config    = cfg.load('./config.json')
+var config    = cfg.load('./neoscan/neoscan.config.json')
 const cmc     = require('./get_cmc_price.js')
 const binance = require('./binance.js')
 
@@ -45,5 +45,5 @@ if (program.debug) {
 
 
 get_price(program.symbol).then(result => {
-  print('usd value: '+ result + '\n' + 'net worth for amount: ' + result * program.amount)
+  if(result) print('usd value: '+ result + '\n' + 'net worth for amount: ' + result * program.amount)
 })
