+++ src/nodejs/neoscan/get_height.js
@@ -3,11 +3,6 @@
 const neoscan = require('./neoscan.js')
 const dbg   = require('../debug')
 const program = require('commander');
-var cfg     = require('./config.json')
-
-var apiv1 = cfg.api.v1;
-
-// dbg.logDeep('apiv1', apiv1)
 
 function print(msg) {
   console.log(msg);
@@ -21,7 +16,6 @@ program
   .parse(process.argv);
 
 if (!program.net) {
-  // print('network: ' + program.net);
 }
 
 if (program.debug) {
