+++ src/nodejs/neoscan/neoscan.config.json
@@ -1,15 +1,4 @@
 {
-  "accounts": {
-    "external": {
-      "default": true,
-      "path": "/home/fet/nwd/phetter/configs/neotools.config.json"
-    },
-    "account_name2": {
-      "address": "",
-      "default": false,
-      "path": ""
-    }
-  },
   "networks": {
     "MainNet": {
       "rootUrl": "https://neoscan-testnet.io/api/test_net/"
